// web/src/proto/database/database_pb.ts
// @generated by protoc-gen-es v2.11.0 with parameter "target=ts,import_extension=ts,json_types=true,valid_types=legacy_required+protovalidate_required"
// @generated from file database/database.proto (package database, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Postgres_Cluster, Postgres_ClusterJson, Postgres_Instance, Postgres_InstanceJson } from "./postgres_pb.ts";
import { file_database_postgres } from "./postgres_pb.ts";
import { file_validate_validate } from "../validate/validate_pb.ts";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file database/database.proto.
 */
export const file_database_database: GenFile = /*@__PURE__*/
  fileDesc("ChdkYXRhYmFzZS9kYXRhYmFzZS5wcm90bxIIZGF0YWJhc2Ui0wEKCERhdGFiYXNlEjcKCHRlbXBsYXRlGAEgASgLMhsuZGF0YWJhc2UuRGF0YWJhc2UuVGVtcGxhdGVCCPpCBYoBAhABGo0BCghUZW1wbGF0ZRI4ChFwb3N0Z3Jlc19pbnN0YW5jZRgBIAEoCzIbLmRhdGFiYXNlLlBvc3RncmVzLkluc3RhbmNlSAASNgoQcG9zdGdyZXNfY2x1c3RlchgCIAEoCzIaLmRhdGFiYXNlLlBvc3RncmVzLkNsdXN0ZXJIAEIPCgh0ZW1wbGF0ZRID+EIBQkBaPmdpdGh1Yi5jb20vc3Ryb3BweS1pby9oYXRjaGV0LXdvcmtmbG93L2ludGVybmFsL3Byb3RvL2RhdGFiYXNlYgZwcm90bzM", [file_database_postgres, file_validate_validate]);

/**
 * @generated from message database.Database
 */
export type Database = Message<"database.Database"> & {
  /**
   * @generated from field: database.Database.Template template = 1;
   */
  template?: Database_Template;
};

/**
 * @generated from message database.Database
 */
export type DatabaseJson = {
  /**
   * @generated from field: database.Database.Template template = 1;
   */
  template?: Database_TemplateJson;
};

export type DatabaseValid = Database;

/**
 * Describes the message database.Database.
 * Use `create(DatabaseSchema)` to create a new message.
 */
export const DatabaseSchema: GenMessage<Database, {jsonType: DatabaseJson, validType: DatabaseValid}> = /*@__PURE__*/
  messageDesc(file_database_database, 0);

/**
 * @generated from message database.Database.Template
 */
export type Database_Template = Message<"database.Database.Template"> & {
  /**
   * @generated from oneof database.Database.Template.template
   */
  template: {
    /**
     * @generated from field: database.Postgres.Instance postgres_instance = 1;
     */
    value: Postgres_Instance;
    case: "postgresInstance";
  } | {
    /**
     * @generated from field: database.Postgres.Cluster postgres_cluster = 2;
     */
    value: Postgres_Cluster;
    case: "postgresCluster";
  } | { case: undefined; value?: undefined };
};

/**
 * @generated from message database.Database.Template
 */
export type Database_TemplateJson = {
  /**
   * @generated from field: database.Postgres.Instance postgres_instance = 1;
   */
  postgresInstance?: Postgres_InstanceJson;

  /**
   * @generated from field: database.Postgres.Cluster postgres_cluster = 2;
   */
  postgresCluster?: Postgres_ClusterJson;
};

export type Database_TemplateValid = Database_Template;

/**
 * Describes the message database.Database.Template.
 * Use `create(Database_TemplateSchema)` to create a new message.
 */
export const Database_TemplateSchema: GenMessage<Database_Template, {jsonType: Database_TemplateJson, validType: Database_TemplateValid}> = /*@__PURE__*/
  messageDesc(file_database_database, 0, 0);